Move couch design info service to libs
