+++ src/services/monitor.ts
@@ -1,7 +1,7 @@
 import * as Effect from 'effect/Effect';
 import * as Context from 'effect/Context';
 import { CouchDbInfo, getDbsInfoByName } from '../libs/couch/dbs-info';
-import { CouchDesignInfo, getDesignInfo } from './couch/design-info';
+import { CouchDesignInfo, getDesignInfo } from '../libs/couch/design-info';
 import { CouchNodeSystem, CouchNodeSystemService } from './couch/node-system';
 import { Array, Clock, Number, Option, pipe } from 'effect';
 import { LocalDiskUsageService } from './local-disk-usage';
